server/requirements.txt
PyJWT==2.8.0
bcrypt==4.1.2
SQLAlchemy==2.0.25
requests==2.31.0
psycopg2==2.9.9
eventlet==0.35.0
python-dotenv==1.0.1
fastapi==0.110.0
python-multipart==0.0.9
uvicorn[standard]==0.27.1
minio==7.2.5
python-jose[cryptography]==3.3.0
apscheduler==3.10.4